requirements.txt
py-cord==2.6.1
python-dotenv==1.0.0
PyNaCl==1.5.0
audioop-lts==0.2.1